Backend/AnomalyDetection/requirements.txt
pymongo>=4.0
python-dotenv>=0.19
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
mlflow>=2.0.0
